+++ sr-tee-io/Cargo.toml
@@ -55,6 +55,9 @@ git = "https://github.com/paritytech/substrate"
 package = "substrate-externalities"
 rev = "3bf9540e72df5ecb3955845764dfee7dcdbb26b5"
 
+[dependencies.sgx_externalities]
+path = "../sgx_externalities"
+
 [features]
 default = ["std"]
 std = [
